fix(profile,org): remove the MFA and RBAC command families that never worked
Every API-backed 'ankra profile auth' command (status, totp, recovery
codes, passkey list/remove) called /api/v1/org/account/mfa*, and the
'ankra org cluster-groups / assign / assignments / unassign / roles
create' family called /api/v1/org/organisation/*: routes the backend only
serves to browser sessions, never on the token lane the CLI authenticates
with. Both families shipped in v0.6.0 and every invocation since has
failed with a 404/405/422.

Two-factor settings are a browser flow by design — enrollment is
session-authenticated and passkeys need a WebAuthn ceremony a terminal
cannot run — so instead of growing bearer-token twins for them, the CLI
now ships a single 'ankra profile auth open' that lands the user on
Profile Authentication. The previously working 'profile auth passkeys
open' forwards there with a deprecation warning (removal v0.10.0).

The shared CSRF JSON helpers move from account_mfa.go to csrf.go (the
managed-clusters lane still uses them), sendJSON moves next to getJSON in
helpers.go, and ListOrganisationUsers survives in org_users.go — it hits
a real route and documents the correct users-array decode shape.

'ankra org roles' (the static list of assignable roles), members, invite,
and every other org command are unaffected.

Bead: ankra-u2eb
